Add files via upload
This project objective is to enable users to be able to know and track the cost of the projects he/she is executing. This folder contains android manifest xml file, houseprojecttracker folder which contains the back end source codes for the application and res folder which it the resource folder for the application and contains the source codes for the user interfaces.